Update made on BookSearchBar, to be absolute
The changes we made in order to prevent pushing the search list.
Both BookSearchBar and BookSearchList are now absolute and are wrapped
in a div that is relative.
Still needs to be worked on the SearchLocation and SearchInput so they
also will be made absolute. In that way we will avoid similar issues on
small screens.